fix(clippy): complexity related fixes

+++ server/graphql/general/src/queries/requisition_line_chart.rs
@@ -115,7 +115,7 @@ impl ConsumptionOptionsInput {
         default_amc_lookback_months: f64,
         from: Option<Self>,
     ) -> ConsumptionHistoryOptions {
-        let default_amc_lookback_months = default_amc_lookback_months as f64;
+        let default_amc_lookback_months = default_amc_lookback_months;
         let default_number_of_datapoints = 3;
 
         from.map(
